requirements.txt
aiohttp
botbuilder-core 
requests 
python-dotenv
gunicorn